REFACTOR: Minor refactoring, Add additional documentation + Examples
