+++ .editorconfig
@@ -1,0 +1,3 @@
+[rs]
+indent_style = space
+indent_size = 4
