+++ Cargo.toml
@@ -18,6 +18,8 @@ path = "src/main.rs"
 
 [dependencies]
 zentime-rs-timer = { path = "./timer", version = "0.8" }
+log = "0.4"
+env_logger = "0.10"
 daemonize = { version = "0.4" }
 interprocess = { version = "1.2", features = ["tokio_support", "signals"]}
 tokio = { version = "1", features = ["full"] }
