+++ src/client.rs
@@ -1,6 +1,7 @@
+//! Code related to zentime terminal clients (e.g. async connection handling, terminal io etc.)
+
 mod connection;
 pub mod start;
 pub mod terminal_io;
-pub mod view;
 
 pub use start::start;
