+++ src/client/start.rs
@@ -1,8 +1,12 @@
+//! Module containing code relevant to starting a zentime terminal client.
+//! See [`start`]
+
+use crate::client::terminal_io::output::RawInterface;
+use crate::client::terminal_io::output::TerminalOutputTask;
 use std::sync::Arc;
 
 use crate::client::terminal_io::input::TerminalInputTask;
 use crate::client::terminal_io::output::TerminalOut;
-use crate::client::terminal_io::terminal::Terminal;
 use crate::config::Config;
 use futures::future::FutureExt;
 use futures::lock::Mutex;
@@ -13,6 +17,27 @@ use super::connection::ClientConnectionTask;
 use crate::client::terminal_io::output::DefaultInterface;
 use crate::client::terminal_io::output::MinimalInterface;
 
+/// Start a single zentime client and connect it to the zentime server.
+/// This makes sure we have tokio tasks in place to:
+/// * listen to incoming input events
+/// * render output to the terminal
+/// * hanndle communication over a client-server connection via IPC-message passing
+///
+/// # Example
+///
+/// ```no_run
+/// use zentime_rs::client::start;
+/// use zentime_rs::config::create_config;
+/// use zentime_rs::config::Config;
+///
+/// #[tokio::main]
+/// async fn main() {
+///     let config: Config = create_config("./some/path/config.toml")
+///        .extract()
+///        .expect("Could not create config");
+///     start(config).await;
+/// }
+/// ```
 pub async fn start(config: Config) {
     let (terminal_in_tx, terminal_in_rx) = unbounded_channel();
     let (terminal_out_tx, terminal_out_rx) = unbounded_channel();
@@ -24,7 +49,7 @@ pub async fn start(config: Config) {
     let thread_safe_terminal_out = Arc::new(Mutex::new(terminal_out));
 
     let input_handler = TerminalInputTask::spawn(terminal_in_tx);
-    let view_handler = Terminal::spawn_renderer(thread_safe_terminal_out.clone(), terminal_out_rx);
+    let view_handler = TerminalOutputTask::spawn(thread_safe_terminal_out.clone(), terminal_out_rx);
     let connection_handler = ClientConnectionTask::spawn(terminal_in_rx, terminal_out_tx);
 
     let join_result = try_join! {
@@ -41,21 +66,26 @@ pub async fn start(config: Config) {
     }
 }
 
+/// Determine which terminal interface should be used.
 fn init_interface(interface_type: String) -> Box<dyn TerminalOut + Send> {
-    if interface_type == "minimal" {
-        // TODO maybe try to avoid duplication between both interface types
-        match MinimalInterface::new() {
+    match interface_type.as_str() {
+        "minimal" => match MinimalInterface::new() {
             Ok(interface) => Box::new(interface),
             Err(error) => {
                 panic!("Could not initialize interface: {}", error);
             }
-        }
-    } else {
-        match DefaultInterface::new() {
+        },
+        "raw" => match RawInterface::new() {
             Ok(interface) => Box::new(interface),
             Err(error) => {
                 panic!("Could not initialize interface: {}", error);
             }
-        }
+        },
+        _ => match DefaultInterface::new() {
+            Ok(interface) => Box::new(interface),
+            Err(error) => {
+                panic!("Could not initialize interface: {}", error);
+            }
+        },
     }
 }
