+++ src/client/terminal_io.rs
@@ -1,4 +1,6 @@
+//! Zentime client terminal io
+
+mod default_interface;
 pub mod input;
 pub mod output;
-pub mod terminal;
 pub mod terminal_event;
