+++ src/server.rs
@@ -1,7 +1,9 @@
+//! Zentime server utilities
+
 pub mod notification;
 mod sound;
 mod start;
+pub mod status;
 mod timer_output;
-pub mod util;
 
 pub use start::start;
