+++ src/server/start.rs
@@ -7,6 +7,7 @@ use crate::server::timer_output::TimerOutputAction;
 use anyhow::Context;
 use crossbeam::channel::{unbounded, Sender};
 use interprocess::local_socket::tokio::OwnedWriteHalf;
+use log::info;
 use tokio::task::{spawn_blocking, yield_now};
 
 use std::sync::Arc;
@@ -21,11 +22,10 @@ use tokio::fs::{metadata, remove_file};
 
 use zentime_rs_timer::{Timer, TimerInputAction};
 
-use super::util::{server_status, ServerStatus};
+use super::status::{server_status, ServerStatus};
 
-// TODO
-// add logging
-
+/// Starts the server by opening the zentime socket and listening for incoming connections.
+/// This will just quit if another zentime server process is already running.
 #[tokio::main]
 pub async fn start(config: Config) -> anyhow::Result<()> {
     let socket_name = get_socket_name();
@@ -33,11 +33,14 @@ pub async fn start(config: Config) -> anyhow::Result<()> {
     let socket_file_already_exists = metadata(socket_name).await.is_ok();
 
     if socket_file_already_exists && server_status() == ServerStatus::Running {
+        info!("Server is already running. Terminating this process...");
         // Apparently a server is already running and we don't need to do anything
         return Ok(());
     }
 
     if socket_file_already_exists {
+        info!("Socket file already exists - removing file");
+
         // We have a dangling socket file without an attached server process.
         // In that case we simply remove the file and start a new server process
         remove_file(socket_name)
@@ -45,14 +48,20 @@ pub async fn start(config: Config) -> anyhow::Result<()> {
             .context("Could not remove existing socket file")?
     };
 
-    if let Err(error) = listen(config, socket_name).await {
-        panic!("Could not start server listener: {}", error);
-    };
+    info!("Start listening for connections...");
+
+    listen(config, socket_name)
+        .await
+        .context("Error while listening for connections")?;
 
     Ok(())
 }
 
+/// This starts a blocking tokio task which runs the actual synchronous timer logic, but
+/// also listens for incoming client connections and spawns a new async task for each incoming
+/// connection.
 async fn listen(config: Config, socket_name: &str) -> anyhow::Result<()> {
+    info!("Binding to socket...");
     let listener =
         LocalSocketListener::bind(socket_name).context("Could not bind to local socket")?;
 
@@ -66,6 +75,8 @@ async fn listen(config: Config, socket_name: &str) -> anyhow::Result<()> {
     let timer_out_tx = timer_output_sender.clone();
 
     spawn_blocking(move || {
+        info!("Starting timer...");
+
         Timer::new(
             config.timers,
             Box::new(move |_, msg| {
@@ -100,6 +111,7 @@ async fn listen(config: Config, socket_name: &str) -> anyhow::Result<()> {
         // and hand the connection over to them so that multiple clients
         // could be processed simultaneously in a lightweight fashion.
         tokio::spawn(async move {
+            info!("New connection received.");
             if let Err(error) = handle_conn(connection, input_tx, output_rx).await {
                 panic!("Could not handle connection: {}", error);
             };
@@ -107,7 +119,9 @@ async fn listen(config: Config, socket_name: &str) -> anyhow::Result<()> {
     }
 }
 
-// Describe the things we do when we've got a connection ready.
+/// Describe the things we do when we've got a connection ready.
+/// This will continously send the current timer state to the client and also listen for incoming
+/// [ClientToServerMsg]s.
 async fn handle_conn(
     conn: LocalSocketStream,
     timer_input_sender: Sender<TimerInputAction>,
@@ -137,6 +151,7 @@ async fn handle_conn(
         yield_now().await;
     }
 
+    info!("Closing connection");
     Ok(())
 }
 
@@ -150,22 +165,37 @@ async fn handle_client_to_server_msg(
     timer_input_sender: &Sender<TimerInputAction>,
 ) -> anyhow::Result<CloseConnection> {
     match msg {
+        // Shutdown server
         ClientToServerMsg::Quit => {
-            println!("\nClient told server to shutdown");
-            println!("Shutting down...");
+            info!("\nClient told server to shutdown");
+
+            info!("Cleaning up socket file");
+            let socket_name = get_socket_name();
+            remove_file(socket_name)
+                .await
+                .context("Could not remove existing socket file")?;
+
+            info!("Shutting down...");
             std::process::exit(0);
         }
+
+        // Play/Pause the timer
         ClientToServerMsg::PlayPause => {
             timer_input_sender
                 .send(TimerInputAction::PlayPause)
                 .context("Could not send Play/Pause to timer")?;
         }
+
+        // Skip to next timer interval
         ClientToServerMsg::Skip => {
             timer_input_sender
                 .send(TimerInputAction::Skip)
                 .context("Could not send Skip to timer")?;
         }
+
+        // Close connection, because client has detached
         ClientToServerMsg::Detach => {
+            info!("Client detached.");
             return Ok(CloseConnection::Yes);
         }
     }
