+++ src/server/timer_output.rs
@@ -1,6 +1,7 @@
 use serde::{Deserialize, Serialize};
 use zentime_rs_timer::timer::ViewState;
 
+/// Carries the timer state as view state
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum TimerOutputAction {
     Timer(ViewState),
