+++ timer/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "zentime-rs-timer"
-version = "0.8.0"
+version = "0.8.1"
 edition = "2021"
 description = "Pausable productivity timer"
 license = "MIT"
